tensor_aio: Remove TTWU hook
The TTWU hook actually results in worse thermal tracking for the scheduler
because it attributes the instantaneous TMU throttle to an entire PELT
window (1024 us). As a result, the throttle measured at a single point in
time is assumed by the scheduler to have been present across the entire
duration of a PELT window.

Since updating the thermal pressure as often as possible is
counterproductive, remove the TTWU hook outright.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>