tensor_aio: Implement CPU hardware thermal throttle detection
Tensor CPUs experience some unknown, invisible CPU hardware thermal
throttling that isn't reported through any software interface. This is
observed via calculating the CPU's real frequency using PMU/AMU counters.

This doesn't appear to be a glitch in either the CPU cycle counter or the
system timer, because microbenching reveals that the CPU is indeed running
slower when the invisible throttling is observed via counter measurements.

The degree to which the CPU is throttled by this mechanism can be quite
extreme (>1000 MHz), and the throttle can persist for several milliseconds,
or even seconds during a heavy sustained workload such as hackbench.

When the scheduler is unaware that the CPU is thermal throttled, it can
make suboptimal scheduling decisions under the assumption that the CPU's
frequency can be scaled up when it really cannot.

To mitigate poor scheduling decisions stemming from the invisible hardware
throttling, implement a mechanism to detect and report it in Tensor AIO.
This is achieved by tracking the actual CPU frequency voted through ACPM
and checking if the measured CPU frequency is below that by a margin which
is significant enough to cause poor scheduling decisions.

This conveniently makes it possible to detect CPU frequency transition
delays that are long enough to impact scheduling, too.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>